Updates to stylsheet, work pg, index page
There were 2 'section1' classes on the stylesheet, removed one.
Removed the font-family in the intro, will explore adding fonts in DW and change the font family.

Removed the <br> syntax on line 61 of the index page.

Work page, highlighted the link 'Here!' so users can interact with it and understand it is a responsive element.